+++ android-client/app/src/main/java/im/agent/personal/AgentNotificationService.kt
@@ -1,4 +1,4 @@
-package im.agent.personal
+package im.agent.link
 
 import android.app.Notification
 import android.app.Service