+++ android-client/app/src/main/java/im/agent/personal/HubConfig.kt
@@ -1,4 +1,4 @@
-package im.agent.personal
+package im.agent.link
 
 import android.app.Application
 import android.os.Build